Back to school and go on working
in page: calendar, we would gonna send the request of getting  todos from api repeatly because request was on the calendar months' switch action. so We set a memory-cache and it'd destroyed when user left the calendar page. it can quitely cancel some unnecessary api request in order to avoid server crash.